requirements.txt
Autologging==1.3.2
python-json-logger==2.0.4
psycopg2==2.9.3
pandas==1.3.5
numpy==1.21.6
PyYAML